added database(sqlite3)
Previously, all data was saved in a .txt file and this made it
impossible for several people to use the bot at the same time.

To do this, a database (sqlite3) was added to the bot, now as many
people as possible can use the bot at the same time.